Create HC-05_Bluetooth_Module_Solution
PROBLEM WITH BLUETOOTH COMUNICATION
For some reason the Arduino constantly receives '-1' from my android (even when I am not sending anything!), so I have to start the void loop() by making the program ignore any inputs with the value -1. Apart from this, whenever I send a one-digit number to the Arduino, it gets received at the end of an chain of 10 constants, and then followed by a 35. So in order to tell the Arduino only to registrate the number I sent I have to count up to 11 whenever a new "chain" is received and when the counter hits 11 take that current received value and set as sentVal. Then when the following "35" is received the counter resets.